+++ Assets/Libraries/Entitas.Unity.VisualDebugging/Entity/EntityBehaviour.cs
@@ -1,21 +1,31 @@
-﻿using UnityEngine;
+﻿using System;
 using Entitas;
+using UnityEngine;
 
 namespace Entitas.Unity.VisualDebugging {
     public class EntityBehaviour : MonoBehaviour {
         public Pool pool { get { return _pool; } }
         public Entity entity { get { return _entity; } }
+        public string[] componentNames { get { return _componentNames; } }
+        public Type[] componentTypes { get { return _componentTypes; } }
         public bool[] unfoldedComponents { get { return _unfoldedComponents; } }
+
+        public int componentToAdd;
 
         Pool _pool;
         Entity _entity;
+        string[] _componentNames;
+        Type[] _componentTypes;
         bool[] _unfoldedComponents;
 
-        public void Init(Pool pool, Entity entity) {
+        public void Init(Pool pool, Entity entity, string[] componentNames, Type[] componentTypes) {
             _pool = pool;
             _entity = entity;
+            _componentNames = componentNames;
+            _componentTypes = componentTypes;
             _unfoldedComponents = new bool[_pool.totalComponents];
-            _entity.OnComponentRemoved += onComponentRemoved;
+            _entity.OnEntityReleased += onEntityReleased;
+            Update();
 
             UnfoldAllComponents();
         }
@@ -33,14 +43,12 @@ namespace Entitas.Unity.VisualDebugging {
         }
 
         public void DestroyBehaviour() {
-            _entity.OnComponentRemoved -= onComponentRemoved;
+            _entity.OnEntityReleased -= onEntityReleased;
             Destroy(gameObject);
         }
 
-        void onComponentRemoved(Entity e, int index, IComponent component) {
-            if (!_pool.HasEntity(e)) {
-                DestroyBehaviour();
-            }
+        void onEntityReleased(Entity e) {
+            DestroyBehaviour();
         }
 
         void Update() {