+++ Assets/Libraries/Entitas.Unity.VisualDebugging/PoolObserver/Editor/PoolObserverEditor.cs
@@ -12,18 +12,28 @@ namespace Entitas.Unity.VisualDebugging {
 
             EditorGUILayout.BeginVertical(GUI.skin.box);
             EditorGUILayout.LabelField(poolObserver.name, EditorStyles.boldLabel);
-            EditorGUILayout.LabelField("Entities", poolObserver.pool.Count.ToString());
+            EditorGUILayout.LabelField("Entities", poolObserver.pool.count.ToString());
             EditorGUILayout.LabelField("Reusable entities", poolObserver.pool.reusableEntitiesCount.ToString());
+            EditorGUILayout.LabelField("Retained entities", poolObserver.pool.retainedEntitiesCount.ToString());
             EditorGUILayout.EndVertical();
+
+            if (GUILayout.Button("Create Entity")) {
+                var creationIndex = poolObserver.pool.CreateEntity().creationIndex;
+                var entityBehaviourName = "Entity_" + creationIndex + "()";
+                var entityBehaviour = Object.FindObjectsOfType<EntityBehaviour>()
+                    .Single(eb => eb.name == entityBehaviourName);
+
+                Selection.activeGameObject = entityBehaviour.gameObject;
+            }
 
             var groups = poolObserver.groups;
             if (groups.Length != 0) {
                 EditorGUILayout.BeginVertical(GUI.skin.box);
                 EditorGUILayout.LabelField("Groups (" + groups.Length + ")", EditorStyles.boldLabel);
-                foreach (var group in groups.OrderByDescending(g => g.Count)) {
+                foreach (var group in groups.OrderByDescending(g => g.count)) {
                     EditorGUILayout.BeginHorizontal();
                     EditorGUILayout.LabelField(group.ToString());
-                    EditorGUILayout.LabelField(group.Count.ToString(), GUILayout.Width(48));
+                    EditorGUILayout.LabelField(group.count.ToString(), GUILayout.Width(48));
                     EditorGUILayout.EndHorizontal();
                 }
                 EditorGUILayout.EndVertical();