+++ Assets/Libraries/Entitas/AffectingSystemsComponent.cs
@@ -1,7 +1,0 @@
-﻿using System.Collections.Generic;
-using Entitas;
-
-public class AffectingSystemsComponent : IComponent
-{
-    public List<string> Systems;
-}