+++ Assets/Libraries/Entitas/Pool.cs
@@ -13,8 +13,9 @@ namespace Entitas {
         public delegate void GroupChanged(Pool pool, Group group);
 
         public int totalComponents { get { return _totalComponents; } }
-        public int Count { get { return _entities.Count; } }
+        public int count { get { return _entities.Count; } }
         public int reusableEntitiesCount { get { return _reusableEntities.Count; } }
+        public int retainedEntitiesCount { get { return _retainedEntities.Count; } }
 
         protected readonly HashSet<Entity> _entities = new HashSet<Entity>(EntityEqualityComparer.comparer);
         protected readonly Dictionary<IMatcher, Group> _groups = new Dictionary<IMatcher, Group>();