+++ Assets/Libraries/Entitas/ReactiveSystem.cs
@@ -8,6 +8,7 @@ namespace Entitas {
         readonly GroupObserver _observer;
         readonly IMatcher _ensureComponents;
         readonly IMatcher _excludeComponents;
+        readonly bool _clearAfterExecute;
         readonly List<Entity> _buffer;
 
         public ReactiveSystem(Pool pool, IReactiveSystem subSystem) :
@@ -29,6 +30,8 @@ namespace Entitas {
                 _excludeComponents = excludeComponents.excludeComponents;
             }
 
+            _clearAfterExecute = (subSystem as IClearReactiveSystem) != null;
+
             var triggersLength = triggers.Length;
             var groups = new Group[triggersLength];
             var eventTypes = new GroupEventType[triggersLength];
@@ -47,6 +50,10 @@ namespace Entitas {
 
         public void Deactivate() {
             _observer.Deactivate();
+        }
+
+        public void Clear() {
+            _observer.ClearCollectedEntities();
         }
 
         public void Execute() {
@@ -79,31 +86,15 @@ namespace Entitas {
 
                 _observer.ClearCollectedEntities();
                 if (_buffer.Count != 0) {
-                    UpdateAffectingSystemsOnEntities();
-
                     _subsystem.Execute(_buffer);
-                    for (int i = 0, bufferCount = _buffer.Count; i < bufferCount; i++)
-                    {
+                    for (int i = 0, bufferCount = _buffer.Count; i < bufferCount; i++) {
                         _buffer[i].Release();
                     }
                     _buffer.Clear();
+                    if (_clearAfterExecute) {
+                        _observer.ClearCollectedEntities();
+                    }
                 }
-            }
-        }
-
-        private void UpdateAffectingSystemsOnEntities()
-        {
-            foreach (var entity in _buffer)
-            {
-                var currentSystems = entity.hasAffectingSystems
-                    ? entity.affectingSystems.Systems
-                    : new List<string>();
-                var systemName = _subsystem.GetType().Name;
-                if (!currentSystems.Contains(systemName))
-                {
-                    currentSystems.Add(systemName);
-                }
-                entity.ReplaceAffectingSystems(currentSystems);
             }
         }
     }